add container for each sections. layout navigation-bar
limit the width of the container of section to 960px which fit most of the desktop screen size and center it to the center of section.

every section will have the same padding and margin so the website can be uniform.

style navigation bar so it can be at the top of the web page where logo on one side and navigation link on the other site. easily to follow